test: 기능, 검증 테스트 분리
